+++ app/build.gradle
@@ -8,7 +8,7 @@ android {
         targetSdkVersion 26
         versionCode 1
         versionName "1.0"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner "android.support.TrackListActivity.runner.AndroidJUnitRunner"
     }
     buildTypes {
         release {
@@ -25,8 +25,8 @@ dependencies {
     implementation 'com.google.firebase:firebase-auth:12.0.0'
     implementation 'com.android.support:design:26.1.0'
     testImplementation 'junit:junit:4.12'
-    androidTestImplementation 'com.android.support.test:runner:1.0.1'
-    androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.1'
+    androidTestImplementation 'com.android.support.TrackListActivity:runner:1.0.1'
+    androidTestImplementation 'com.android.support.TrackListActivity.espresso:espresso-core:3.0.1'
     implementation 'com.google.firebase:firebase-core:12.0.0'
     implementation 'com.google.firebase:firebase-database:12.0.0'
     implementation 'com.google.zxing:core:3.2.1'
