--- app/src/main/java/com/parcelpaladin/d30f2n/parcelpaladinfirebase/TrackingInfo.java
@@ -1,18 +1,0 @@
-package com.parcelpaladin.d30f2n.parcelpaladinfirebase;
-
-/**
- * Created by d30f2n on 3/20/2018.
- */
-
-public class TrackingInfo {
-    public String tracking;
-
-    public TrackingInfo()
-    {
-
-    }
-
-    public TrackingInfo(String tracking) {
-        this.tracking = tracking;
-    }
-}
