--- app/src/main/res/layout/activity_profile.xml
@@ -1,69 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context="com.parcelpaladin.d30f2n.parcelpaladinfirebase.ProfileActivity">
-
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_marginTop="50dp"
-        android:orientation="vertical"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.0"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="1.0">
-
-        <TextView
-            android:id="@+id/textViewUserEmail"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="center_horizontal"
-            android:text="Welcome: name"
-            android:textAppearance="@style/TextAppearance.AppCompat" />
-
-
-        <Space
-            android:layout_width="match_parent"
-            android:layout_height="75dp" />
-
-        <EditText
-            android:id="@+id/editTextTracking"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_gravity="center_horizontal"
-            android:ems="10"
-            android:hint="Enter your tracking number here"
-            android:inputType="none" />
-
-        <Button
-            android:id="@+id/buttonSave"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:text="Save" />
-
-        <Button
-            android:id="@+id/buttonTracking"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:text="Show Tracking" />
-
-        <Space
-            android:layout_width="match_parent"
-            android:layout_height="171dp" />
-
-        <!--<Button-->
-            <!--android:id="@+id/buttonLogout"-->
-            <!--android:layout_width="wrap_content"-->
-            <!--android:layout_height="wrap_content"-->
-            <!--android:layout_gravity="center_horizontal"-->
-            <!--android:text="Log out" />-->
-
-
-    </LinearLayout>
-</android.support.constraint.ConstraintLayout>
